Reorganisation and midway commit for FileHandlerTest
